\name{asOneFormula}
\alias{allVarsRec}
\alias{asOneFormula}

\title{Combine Formulas of a Set of Objects}

\description{
Auxiliary functions to combine formulas of a set of objects. See \code{help("asOneFormula")} from package \code{nlme}.
}

\source{
Jose Pinheiro, Douglas Bates, Saikat DebRoy, Deepayan Sarkar and the R Development Core Team (2011). nlme: Linear and Nonlinear Mixed Effects Models. R package version 3.1-100. \url{http://CRAN.R-project.org/package=nlme}
}

\author{Original version by Jose Pinheiro and Douglas Bates}
