\name{cov.sel}
\alias{covHandling}
\alias{errorHandling}
\alias{extractAll}
\alias{loglik.s}
\alias{loglik.t}
\alias{quad}
\alias{theta.z.dim}
\alias{cov.sel}

\title{
Accessory Functions
}
\description{
Undocumented functions
}

\usage{
cov.sel(type)
}

\arguments{
  \item{type}{
...
}
}

\author{
Marco Geraci
}