\name{F.lqm}
\alias{F.lqm}

\title{
Accessory Function
}

\description{
Accessory function for \code{\link{lqm.counts}}.
}
\usage{
F.lqm(x, cn)
}

\arguments{
  \item{x}{
value.
}
  \item{cn}{
small constant.
}
}

\details{
This function is used by \code{\link{lqm.counts}}. See Theorem 3 in Machado and Santos Silva (2005).
}


\references{
Machado JAF and Santos Silva JMC (2005). Quantiles for counts. Journal of the American Statistical Association, 100(472), 1226--1237. 
}

\author{
Marco Geraci
}

\seealso{
\code{\link{lqm.counts}}
}

\keyword{counts}
