\name{gauss.quad.prob}
\alias{gauss.quad.prob}

\title{
Gaussian Quadrature
}

\description{
This function calculates nodes and weights for Gaussian quadrature in terms of probability distributions. See \code{help("gauss.quad.prob")} from package \code{statmod}.
}

\source{
Gordon Smyth with contributions from Yifang Hu, Peter Dunn and Belinda Phipson. (2011). statmod: Statistical Modeling. R package version 1.4.11. \url{https://CRAN.R-project.org/package=statmod}
}

\author{
Original version by Gordon Smyth
}

\keyword{gaussian quadrature}
