\name{logLik.lqm}
\alias{logLik.lqm}

\title{
Extract Log-Likelihood
}
\description{
\code{logLik.lqm} extracts the log-likelihood of a fitted LQM.
}
\usage{
\method{logLik}{lqm}(object, ...)
}

\arguments{
  \item{object}{
an object of \code{\link{class}} "lqm".

}
  \item{\dots}{
not used.
}
}


\author{
Marco Geraci
}


\seealso{
\code{\link{lqm}}
\code{\link{AIC}}
}

\keyword{models}
