\name{logLik.lqmm}
\alias{logLik.lqmm}

\title{
Extract Log-Likelihood
}
\description{
\code{logLik.lqmm} extracts the log-likelihood of a fitted LQMM.
}
\usage{
\method{logLik}{lqmm}(object, ...)
}

\arguments{
  \item{object}{
an object of \code{\link{class}} "lqmm".

}
  \item{\dots}{
not used.
}
}


\author{
Marco Geraci
}


\seealso{
\code{\link{lqmm}}
\code{\link{AIC}}
}

\keyword{models}
