\name{lqmm-internal}
\title{Internal lqmm objects}
\alias{.First.lib}
\alias{.Last.lib}
\alias{errorHandling}
\alias{extractAll}
\alias{loglik.s}
\alias{loglik.t}
\alias{theta.z.dim}
\alias{createLaguerre}
\alias{quad}
\alias{invTfun}
\alias{Tfun}
\alias{F.lqm}
\alias{addnoise}
\alias{permutations}
\alias{asOneFormula}
\alias{allVarsRec}
\description{Internal lqmm objects.}
\details{These are not to be called by the user.}
\keyword{internal}