\name{lqmm-package}
\alias{lqmm-package}
\docType{package}
\title{
Linear Quantile Models and Linear Quantile Mixed Models}

\description{
Fit quantile regression models for independent and hierarchical data
}

\details{
\tabular{ll}{
Package: \tab lqmm\cr
Type: \tab Package\cr
Version: \tab 1.04\cr
Date: \tab 2014-01-09\cr
License: \tab GPL (>=2)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Marco Geraci

Maintainer: Marco Geraci <m.geraci@ucl.ac.uk>
}

\references{
Geraci M and Bottai M (2007). Quantile regression for longitudinal data using the asymmetric Laplace distribution. Biostatistics 8(1), 140--154.

Geraci M and Bottai M (2013). Linear quantile mixed models. Statistics and Computing, doi: 10.1007/s11222-013-9381-9.

Geraci M. (2014) Linear quantile mixed models: The lqmm package for Laplace quantile regression, Journal of Statistical Software (in press).
}

\keyword{quantile regression}
