\name{make.positive.definite}
\alias{make.positive.definite}

\title{
Compute Nearest Positive Definite Matrix
}
\description{
This function computes the nearest positive definite of a real symmetric matrix.
See \code{help("make.positive.definite")} from package \code{corpcor}.
}

\source{
Juliane Schaefer, Rainer Opgen-Rhein, Verena Zuber, A. Pedro Duarte Silva and Korbinian Strimmer. (2011). corpcor: Efficient Estimation of Covariance and (Partial) Correlation. R package version 1.6.0. \url{https://CRAN.R-project.org/package=corpcor}
}
\author{
Original version by Korbinian Strimmer
}

\keyword{positive definite}
\keyword{covariance}
