\name{permutations}
\alias{permutations}

\title{
Permutations
}

\description{
This function enumerates the possible permutations. See \code{help("permutations")} from package \code{gtools}.
}

\source{
Gregory R. Warnes. Includes R source code and/or documentation contributed by Ben Bolker and Thomas Lumley (2010). gtools: Various R programming tools. R package version 2.6.2. \url{http://CRAN.R-project.org/package=gtools}
}

\author{
Original version by Bill Venables. Extended to handle \code{repeats.allowed} by Gregory R. Warnes
}

\keyword{permutations}
