\name{print.lqm}
\alias{print.lqm}

\title{
Print an \code{lqm} Object
}
\description{
Print an object generated by \code{\link{lqm}}.
}
\usage{
\method{print}{lqm}(x, digits = max(3, getOption("digits") - 3), ...)
}

\arguments{
  \item{x}{
an \code{lqm} object.
}
  \item{digits}{
a non-null value for digits specifies the minimum number of significant digits to be printed in values.
}

  \item{\dots}{
not used.
}
}


\author{
Marco Geraci
}


\seealso{
\code{\link{lqm}}
}

\keyword{print}
