\name{print.lqmm}
\alias{print.lqmm}

\title{
Print an \code{lqmm} Object
}
\description{
Print an object generated by \code{\link{lqmm}}.
}
\usage{
\method{print}{lqmm}(x, digits = max(3, getOption("digits") - 3), ...)
}

\arguments{
  \item{x}{
an \code{lqmm} object.
}
  \item{digits}{
a non-null value for digits specifies the minimum number of significant digits to be printed in values.
}

  \item{\dots}{
not used.
}
}


\author{
Marco Geraci
}


\seealso{
\code{\link{lqmm}}
}

\keyword{print}
