\name{print.summary.lqm}
\alias{print.summary.lqm}

\title{
Print an \code{lqm} Summary Object
}
\description{
Print summary of an \code{lqm} object.}

\usage{
\method{print}{summary.lqm}(x, ...)
}

\arguments{
  \item{x}{
a \code{summary.lqm} object.
}
  \item{\dots}{
not used.
}
}


\author{
Marco Geraci
}

\seealso{
\code{\link{lqm}}, \code{\link{summary.lqm}}
}


\keyword{print}
\keyword{summary}

