\name{print.summary.lqmm}
\alias{print.summary.lqmm}

\title{
Print an \code{lqmm} Summary Object
}
\description{
Print summary of an \code{lqmm} object.}

\usage{
\method{print}{summary.lqmm}(x, digits = max(3, getOption("digits") - 3), ...)
}

\arguments{
  \item{x}{
a \code{summary.lqmm} object.
}
  \item{digits}{
a non-null value for digits specifies the minimum number of significant digits to be printed in values.
}

  \item{\dots}{
not used.
}
}


\author{
Marco Geraci
}

\seealso{
\code{\link{lqmm}}, \code{\link{summary.lqmm}}
}


\keyword{print}
\keyword{summary}

