\name{quad}
\alias{errorHandling}
\alias{extractAll}
\alias{loglik.s}
\alias{loglik.t}
\alias{theta.z.dim}
\alias{covHandling}
\alias{createLaguerre}
\alias{quad}

\title{
Accessory Functions
}
\description{
Undocumented functions
}

\usage{
quad(q, k, type = c("normal","robust"), rule = 1)
}

\arguments{
  \item{q}{
...
}

  \item{k}{
...
}

  \item{type}{
...
}

  \item{rule}{
...
}


}

\author{
Marco Geraci
}