\name{residuals.lqm}
\alias{residuals.lqm}

\title{
Residuals from an \code{lqm} Object
}

\description{
This function computes the residuals from a fitted \code{lqm} object.
}

\usage{
\method{residuals}{lqm}(object, ...)
}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
an \code{lqm} object.}

  \item{\dots}{
not used.
}
}


\value{
a matrix of residuals.
}


\author{
Marco Geraci
}


\seealso{
\code{\link{lqm}}, \code{\link{predict.lqm}}, \code{\link{coef.lqm}}
}

\keyword{residuals}
