\name{Tfun}
\alias{invTfun}
\alias{Tfun}


\title{
Accessory Functions for Duplication Matrices
}
\description{
Undocumented functions
}

\usage{
Tfun(n, type = "pdSymm")
}

\arguments{
  \item{n}{
size of the square matrix.
}
  \item{type}{
one of: \code{pdSymm} general positive--definite matrix, with no additional structure; \code{pdDiag} diagonal; \code{pdIdent} multiple of an identity; \code{pdCompSymm} compound symmetry structure (constant diagonal and constant off--diagonal elements).
}
}

\value{
...
}

\author{
Marco Geraci
}
