\name{lqmm-internal}
\title{Internal lqmm objects}
\alias{.First.lib}
\alias{.Last.lib}
\alias{bandwidth.rq}
\alias{errorHandling}
\alias{extractAll}
\alias{loglik.al}
\alias{loglik.s}
\alias{loglik.t}
\alias{logliki}
\alias{score.al}
\alias{gradientSi}
\alias{switch_check}
\alias{theta.z.dim}
\alias{createLaguerre}
\alias{quad}
\alias{invTfun}
\alias{Tfun}
\alias{F.lqm}
\alias{addnoise}
\alias{permutations}
\alias{asOneFormula}
\alias{allVarsRec}
\alias{nlloglikh}
\alias{rePred}
\alias{C_gradientSh}
\alias{C_gradientSi}
\alias{C_ll_h}
\description{Internal lqmm objects.}
\details{These are not to be called by the user.}
\keyword{internal}