\name{lqmm-package}
\alias{lqmm-package}
\docType{package}
\title{
Linear Quantile Models and Linear Quantile Mixed Models}

\description{
Fit quantile regression models for independent and hierarchical data
}

\details{
\tabular{ll}{
Package: \tab lqmm\cr
Type: \tab Package\cr
Version: \tab 1.02\cr
Date: \tab 2012-08-15\cr
License: \tab GPL (>=3)\cr
LazyLoad: \tab yes\cr
}

}
\author{
Marco Geraci

Maintainer: Marco Geraci <m.geraci@ucl.ac.uk>
}

\references{
Geraci M and Bottai M (2007). Quantile regression for longitudinal data using the asymmetric Laplace distribution. Biostatistics 8(1), 140--154.

Geraci M and Bottai M (1 June 2011). Linear Quantile Mixed Models. Unpublished manuscript.
}

\keyword{quantile regression}
