\name{print.lqm}
\alias{print.lqm}
\alias{print.lqm.counts}

\title{
Print LQM Objects
}
\description{
Print an object generated by \code{\link{lqm}} or \code{\link{lqm.counts}}.
}
\usage{
\method{print}{lqm}(x, digits = max(6, getOption("digits")), ...)
}

\arguments{
  \item{x}{
an \code{lqm} or \code{lqm.counts} object.
}
  \item{digits}{
a non-null value for digits specifies the minimum number of significant digits to be printed in values.
}

  \item{\dots}{
not used.
}
}


\author{
Marco Geraci
}


\seealso{
\code{\link{lqm}}, \code{\link{lqm.counts}}
}

\keyword{print}
